requirements.txt
pygame>=2.0.0
PyOpenGL>=3.1.1
PyOpenGL_accelerate>=3.1.1
